Move tooltip_view into highlights_view
